feat: Implement crypto trading and history retrieval APIs
- Added an API to enable users to trade cryptocurrencies at the latest best aggregated prices. This API does not interface with third-party systems as per the specified requirements.
- Developed an API for users to check their cryptocurrency wallet balances, enhancing the user's ability to manage and monitor their assets.
- Created an API that allows users to view their trading history, providing insights into past transactions and trading patterns.
- Included a README.md file with detailed documentation on the new APIs, usage examples, and configuration instructions for a better developer experience.

Note: All new APIs are designed with mock data constructors for testing purposes, which can be omitted from production as needed.

+++ src/main/java/com/cryptotrading/model/CryptoPrice.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -22,10 +21,10 @@ public class CryptoPrice {
     private String symbol;
 
     @Column(nullable = false)
-    private BigDecimal bidPrice;
+    private double bidPrice;
 
     @Column(nullable = false)
-    private BigDecimal askPrice;
+    private double askPrice;
 
     @Column(name = "price_timestamp", nullable = false)
     private LocalDateTime timestamp;
